requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.0.350
transformers==4.35.2
PyPDF2==3.0.1
faiss-cpu==1.7.4
google-generativeai==0.3.2
langchain-google-genai==0.0.4
langchain-community==0.0.19
scikit-learn==1.3.2
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
